requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asttokens
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
av==14.3.0
backcall
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm
contourpy==1.3.0
cycler==0.12.1
datasets==3.2.0
debugpy
decorator
deepspeed==0.16.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup
executing==2.2.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
git-lfs==1.6
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==7.1.0
importlib_resources==6.5.2
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
-e git+https://github.com/01choco/Dataset-Cartography@0e8479bdd9c52c26a1deb966da558c1e37d2c940#egg=llamafactory&subdirectory=train-eval/LLaMA-Factory
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.58.1
orjson==3.10.16
packaging==24.0
pandas==2.2.3
parso==0.8.3
peft==0.15.1
pexpect
pickleshare
pillow==10.4.0
platformdirs==3.11.0
pooch==1.8.2
prompt-toolkit
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.2
psutil==5.9.8
ptyprocess
pure-eval
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.2.1
python-dateutil
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rouge-chinese==1.0.3
rouge_score==0.1.2
rsa==4.9
ruff==0.11.5
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.26.0
setproctitle==1.3.5
shellingham==1.5.4
shtab==1.7.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.0
tomlkit==0.12.0
torch==2.5.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
transformers==4.51.3
triton==3.1.0
trl==0.9.6
typer==0.15.2
typing_extensions
tyro==0.8.14
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
wandb==0.19.9
wcwidth
websockets==12.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.18.1
